requirements.txt
Django==4.2.2
djangorestframework==3.14.0
coreapi==2.3.3
drf-yasg==1.21.6
psycopg==3.1.9
Pillow==9.5.0